+++ build.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 [ ! -d "build" ]&&mkdir build
-python src/_build/main.py "$@"
+python3 src/_build/main.py "$@"